hayashida

+++ src/main/resources/schema.sql
@@ -1,11 +1,13 @@
 CREATE TABLE IF NOT EXISTS public.product
 (
-    product_id bigint NOT NULL,
+    product_id SERIAL NOT NULL,
     product_name character varying COLLATE pg_catalog."default" NOT NULL,
     product_price integer NOT NULL,
     product_stock integer NOT NULL,
     CONSTRAINT product_pkey PRIMARY KEY (product_id)
 );
+
+
 
 CREATE TABLE IF NOT EXISTS public.reservation
 (
@@ -30,7 +32,7 @@ CREATE TABLE IF NOT EXISTS public.user_table
 -- reservation_product 中間テーブルの作成
 CREATE TABLE IF NOT EXISTS reservation_product (
                                                    reservation_id bigint NOT NULL,
-                                                   product_id bigint NOT NULL,
+                                                   product_id SERIAL NOT NULL,
                                                    FOREIGN KEY (reservation_id) REFERENCES reservation (reservation_id),
                                                    FOREIGN KEY (product_id) REFERENCES product (product_id),
                                                    PRIMARY KEY (reservation_id, product_id)
